Move dehydration logic from the riak backend into the base backend
